Fix an issue where two primitive types couldn't be added to the scene

+++ jme3-core/src/com/jme3/gde/core/sceneexplorer/nodes/primitives/CreateQuadPanel.java
@@ -51,12 +51,12 @@ public class CreateQuadPanel extends javax.swing.JPanel {
         return abstractNewGeometryPanel1;
     }
 
-    public int getQuadWidth() {
-        return (int) spinnerX.getValue();
+    public float getQuadWidth() {
+        return (float) spinnerX.getValue();
     }
 
-    public int getQuadHeight() {
-        return (int) spinnerY.getValue();
+    public float getQuadHeight() {
+        return (float) spinnerY.getValue();
     }
 
     public boolean isFlipCoords() {
